admin gestion

+++ blog/views.py
@@ -8,8 +8,16 @@ from blog.forms import ContactUsForm
 from django.db.models import Q
 from itertools import chain
 from django.core.paginator import Paginator
-
+from django.core.exceptions import PermissionDenied
 from . import forms, models
+from django.contrib.admin.views.decorators import staff_member_required
+from django.contrib.auth import get_user_model
+from .forms import CategoryForm
+
+
+@staff_member_required
+def admin_panel(request):
+    return render(request, 'admin_panel/dashboard.html')
 
 
 @login_required
@@ -38,16 +46,20 @@ def blog_and_photo_upload(request):
             photo = photo_form.save(commit=False)
             photo.uploader = request.user
             photo.save()
+
             blog = blog_form.save(commit=False)
             blog.photo = photo
+            blog.author = request.user 
             blog.save()
+
             blog.contributors.add(request.user, through_defaults={'contribution': 'Auteur principal'})
             return redirect('home')
     context = {
         'blog_form': blog_form,
         'photo_form': photo_form,
-}
+    }
     return render(request, 'blog/create_blog_post.html', context=context)
+
 
 
 
@@ -60,13 +72,57 @@ def view_blog(request, blog_id):
 @login_required
 def view_photo(request, photo_id):
     photo = get_object_or_404(models.Photo, id=photo_id)
-    return render(request, 'blog/view_photo.html', {'photo': photo})
+    comments = photo.comments.all()
+    comment_form = forms.PhotoCommentForm()
 
-# @login_required
-# def home(request):
-#     photos = models.Photo.objects.all()
-#     blogs = models.Blog.objects.all()
-#     return render(request, 'blog/home.html', context={'photos': photos, 'blogs': blogs})
+    if request.method == 'POST':
+        comment_form = forms.PhotoCommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.user = request.user
+            comment.photo = photo
+            comment.save()
+            return redirect('view_photo', photo_id=photo_id)
+
+    return render(request, 'blog/view_photo.html', {
+        'photo': photo,
+        'comments': comments,
+        'comment_form': comment_form
+    })
+
+
+
+@login_required
+def edit_photo(request, photo_id):
+    photo = get_object_or_404(models.Photo, id=photo_id)
+
+    if photo.uploader != request.user:
+        raise PermissionDenied
+
+    form = forms.PhotoForm(instance=photo)
+
+    if request.method == 'POST':
+        form = forms.PhotoForm(request.POST, request.FILES, instance=photo)
+        if form.is_valid():
+            form.save()
+            return redirect('view_photo', photo_id=photo.id)
+
+    return render(request, 'blog/edit_photo.html', {'form': form, 'photo': photo})
+
+@login_required
+def delete_photo(request, photo_id):
+    photo = get_object_or_404(models.Photo, id=photo_id)
+
+    if photo.uploader != request.user:
+        raise PermissionDenied
+
+    if request.method == 'POST':
+        photo.delete()
+        return redirect('home')
+
+    return render(request, 'blog/delete_photo.html', {'photo': photo})
+
+
 
 def home(request):
     follows = request.user.follows.all()  # Récupération des utilisateurs suivis par l'utilisateur actuel
@@ -103,28 +159,44 @@ def home(request):
 
 
 
-
+@login_required
 def photo_feed(request):
+    categorie_id = request.GET.get('categorie')
+    categories = models.Categorie.objects.all()
+
+    # Récupérer les photos des créateurs suivis
     photos = models.Photo.objects.filter(
         uploader__in=request.user.follows.all()
-    ).order_by('-date_created')
+    )
 
-    paginator = Paginator(photos, 6) 
-    page_number = request.GET.get('page')  # Récupérer le numéro de page demandé, par défaut 1
-    page_obj = paginator.get_page(page_number)  # Obtenir l'objet Page correspondant à la page demandée
+    # Si un filtre de catégorie est appliqué
+    if categorie_id:
+        photos = photos.filter(categorie_id=categorie_id)
+
+    # Tri par date décroissante
+    photos = photos.order_by('-date_created')
+
+    # Pagination
+    paginator = Paginator(photos, 6)
+    page_number = request.GET.get('page')
+    page_obj = paginator.get_page(page_number)
 
     context = {
-        'page_obj': page_obj,  # Inclure l'objet Page dans le contexte
+        'page_obj': page_obj,
+        'categories': categories,
     }
 
     return render(request, 'blog/photo_feed.html', context=context)
 
 
-
 @login_required
-@permission_required('blog.change_blog')
 def edit_blog(request, blog_id):
     blog = get_object_or_404(models.Blog, id=blog_id)
+
+    # ✅ Vérifie si l'utilisateur est bien l'auteur
+    if blog.author != request.user:
+        raise PermissionDenied  # Affiche une erreur 403
+
     edit_form = forms.BlogForm(instance=blog)
     delete_form = forms.DeleteBlogForm()
 
@@ -145,6 +217,7 @@ def edit_blog(request, blog_id):
         'delete_form': delete_form,
     }
     return render(request, 'blog/edit_blog.html', context=context)
+
 
 
 @login_required
@@ -178,8 +251,91 @@ def follow_users(request):
 def about(request):
     return render(request, 'blog/about.html')
 
-from django.core.mail import send_mail
-...
+
+# Gérer les utilisateurs
+User = get_user_model()
+
+@staff_member_required
+def admin_users(request):
+    users = User.objects.all()
+
+    if request.method == 'POST':
+        user_id = request.POST.get('user_id')
+        new_role = request.POST.get('new_role')
+        user = get_object_or_404(User, id=user_id)
+        if new_role in [User.CREATOR, User.SUBSCRIBER]:
+            user.role = new_role
+            user.save()
+        return redirect('admin_users')
+
+    return render(request, 'admin_panel/users.html', {'users': users})
+
+@staff_member_required
+def admin_delete_user(request, user_id):
+    user = get_object_or_404(User, id=user_id)
+    if request.user != user:
+        user.delete()
+    return redirect('admin_users')
+
+
+# Gérer les photos
+@staff_member_required
+def admin_photos(request):
+    photos = models.Photo.objects.all()
+    return render(request, 'admin_panel/photos.html', {'photos': photos})
+
+@staff_member_required
+def admin_delete_photo(request, photo_id):
+    photo = get_object_or_404(models.Photo, id=photo_id)
+    photo.delete()
+    return redirect('admin_photos')
+
+# Gérer les blogs
+@staff_member_required
+def admin_blogs(request):
+    blogs = models.Blog.objects.all()
+    return render(request, 'admin_panel/blogs.html', {'blogs': blogs})
+
+@staff_member_required
+def admin_delete_blog(request, blog_id):
+    blog = get_object_or_404(models.Blog, id=blog_id)
+    blog.delete()
+    return redirect('admin_blogs')
+
+# Gérer les catégories
+
+@staff_member_required
+def admin_categories(request):
+    categories = models.Categorie.objects.all()
+    return render(request, 'admin_panel/categories.html', {'categories': categories})
+
+@staff_member_required
+def admin_add_category(request):
+    form = CategoryForm(request.POST or None)
+    if form.is_valid():
+        form.save()
+        return redirect('admin_categories')
+    return render(request, 'admin_panel/category_form.html', {'form': form, 'action': 'Ajouter'})
+
+@staff_member_required
+def admin_edit_category(request, cat_id):
+    category = get_object_or_404(models.Categorie, id=cat_id)
+    form = CategoryForm(request.POST or None, instance=category)
+    if form.is_valid():
+        form.save()
+        return redirect('admin_categories')
+    return render(request, 'admin_panel/category_form.html', {'form': form, 'action': 'Modifier'})
+
+@staff_member_required
+def admin_delete_category(request, cat_id):
+    category = get_object_or_404(models.Categorie, id=cat_id)
+    category.delete()
+    return redirect('admin_categories')
+
+
+
+
+
 
 def contact(request):
     if request.method == 'POST':
@@ -203,3 +359,6 @@ def contact(request):
 
 def confirmation(request):
     return render(request, 'blog/confirmation.html')
+
+
+
